+++ CHANGELOG.md
@@ -4,3 +4,7 @@
 ## 24th December 2022
 
 - [created Cheatsheet files](https://github.com/ftd-lang/ftd/commit/8df76b5b66dd31b9c647a848c6dd4277b434c7fe)
+
+## 26th December 2022
+
+- Added more variants in `ftd.length` (calc, vh, vw, em, rem)